refactor : HandlebarsUtils
