+++ src/main/java/webserver/ExceptionHandler.java
@@ -1,6 +1,6 @@
-package exception;
+package webserver;
 
-import controller.RequestHandler;
+import exception.HttpException;
 import domain.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
