+++ src/main/java/webserver/RequestHandler.java
@@ -1,8 +1,8 @@
-package controller;
+package webserver;
 
+import controller.*;
 import domain.HttpRequest;
 import domain.HttpResponse;
-import exception.ExceptionHandler;
 import exception.HttpException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
