+++ src/main/java/webserver/WebServer.java
@@ -3,7 +3,6 @@ package webserver;
 import java.net.ServerSocket;
 import java.net.Socket;
 
-import controller.RequestHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
